+++ NEWS.md
@@ -1,5 +1,8 @@
 # fitzRoy (development version)
 
+## Bug Fixes
+* Fixed `get_footywire_betting_odds` to return an empty data frame when only future seasons are requested rather than raising an error ([#112](https://github.com/jimmyday12/fitzRoy/issues/112), [@cfranklin11](https://github.com/cfranklin11))
+
 # fitzRoy 0.3.1
 
 * Updating vignettes to use internal data rather than downloading from the internet